+++ Cargo.toml
@@ -95,6 +95,7 @@ fastrand = { path = "libs/fastrand" }
 fiber = { path = "libs/fiber" }
 abort = { path = "libs/abort" }
 panic-unwind = { path = "libs/panic-unwind" }
+util = { path = "libs/util"}
 
 # 3rd-party dependencies
 cfg-if = "1.0.0"
