+++ kernel/src/sync/mod.rs
@@ -5,13 +5,11 @@
 // http://opensource.org/licenses/MIT>, at your option. This file may not be
 // copied, modified, or distributed except according to those terms.
 
-mod cache_padded;
 mod error;
 mod wait_cell;
 mod wait_queue;
 mod wake_batch;
 
-pub use cache_padded::CachePadded;
 pub use error::Closed;
 pub use wait_cell::WaitCell;
 pub use wait_queue::WaitQueue;
