+++ kernel/src/util/mod.rs
@@ -9,7 +9,6 @@ use core::ptr::NonNull;
 
 pub mod atomic_cell;
 pub mod either;
-pub mod maybe_uninit;
 pub mod zip_eq;
 
 /// Helper to construct a `NonNull<T>` from a raw pointer to `T`, with null
