+++ libs/spin/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 [dependencies]
 cfg-if.workspace = true
 lock_api = { workspace = true, optional = true }
+util.workspace = true
 
 [target.'cfg(loom)'.dependencies]
 loom = "0.7"
