+++ libs/spin/src/loom.rs
@@ -93,32 +93,3 @@ cfg_if::cfg_if! {
         }
     }
 }
-
-macro_rules! loom_const_fn {
-    (
-        $(#[$meta:meta])*
-        $vis:vis unsafe fn $name:ident($($arg:ident: $T:ty),*) -> $Ret:ty $body:block
-    ) => {
-        $(#[$meta])*
-        #[cfg(not(loom))]
-        $vis const unsafe fn $name($($arg: $T),*) -> $Ret $body
-
-        $(#[$meta])*
-        #[cfg(loom)]
-        $vis unsafe fn $name($($arg: $T),*) -> $Ret $body
-    };
-    (
-        $(#[$meta:meta])*
-        $vis:vis fn $name:ident($($arg:ident: $T:ty),*) -> $Ret:ty $body:block
-    ) => {
-        $(#[$meta])*
-        #[cfg(not(loom))]
-        $vis const fn $name($($arg: $T),*) -> $Ret $body
-
-        $(#[$meta])*
-        #[cfg(loom)]
-        $vis fn $name($($arg: $T),*) -> $Ret $body
-    }
-}
-
-pub(crate) use loom_const_fn;
