+++ libs/util/src/lib.rs
@@ -1,0 +1,10 @@
+//! Shared utility types & function for k23
+
+#![cfg_attr(not(test), no_std)]
+
+mod cache_padded;
+mod checked_maybe_uninit;
+mod loom;
+
+pub use cache_padded::CachePadded;
+pub use checked_maybe_uninit::{CheckedMaybeUninit, MaybeUninitExt};
